+++ src/main/java/org/folio/circulation/domain/RequestType.java
@@ -3,4 +3,5 @@ package org.folio.circulation.domain;
 public class RequestType {
   public final static String RECALL = "Recall";
   public final static String HOLD = "Hold";
+  public static final String PAGE = "Page";
 }